dermatology_ai_app/requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0

# Enterprise Features - Database & ORM
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
databases>=0.8.0

# Enterprise Features - Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
itsdangerous>=2.1.2

# Enterprise Features - Session & Cache
redis>=5.0.1
python-redis>=5.0.1

# Enterprise Features - Email
fastapi-mail>=1.4.1
email-validator>=2.1.0

# Enterprise Features - Monitoring & Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Enterprise Features - Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Interpretability and Visualization
shap>=0.42.0
lime>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
gradio>=3.40.0

# Image Processing and Analysis
albumentations>=1.3.0
imageio>=2.31.0
scikit-image>=0.21.0

# Database and Storage
sqlalchemy>=2.0.0
sqlite3
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
tqdm>=4.65.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
